🧬 Move the common logic into base module.
