+++ test/unit_test/model/openapi/config.py
@@ -13,7 +13,12 @@ from pymock_api.model import MockAPI
 from pymock_api.model.api_config import _Config
 from pymock_api.model.api_config.apis import APIParameter as PyMockAPIParameter
 from pymock_api.model.enums import OpenAPIVersion, ResponseStrategy
-from pymock_api.model.openapi._base import Transferable, set_openapi_version
+from pymock_api.model.openapi._base import (
+    Transferable,
+    _YamlSchema,
+    set_component_definition,
+    set_openapi_version,
+)
 from pymock_api.model.openapi._parser_factory import (
     BaseOpenAPISchemaParserFactory,
     OpenAPIV2SchemaParserFactory,
@@ -28,9 +33,7 @@ from pymock_api.model.openapi.config import (
     API,
     APIParameter,
     OpenAPIDocumentConfig,
-    _YamlSchema,
     convert_js_type,
-    set_component_definition,
 )
